Added setupPanel with communication to kernel
