+++ src/main/java/ui/debug_panel/setup_panel/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the setup panel component
+ */
+package ui.debug_panel.setup_panel;
